controller: Move mapped namespace up one level
