+++ src/Snowflake.Framework.Primitives/Input/Controller/IControllerElementMappingProfile.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Snowflake.Input.Controller.Mapped
+namespace Snowflake.Input.Controller
 {
     /// <summary>
     /// A collection of mapped elements that describes an input mapping profile.
