+++ src/Snowflake.Framework.Primitives/Input/Device/IDeviceLayoutMapping.cs
@@ -1,5 +1,4 @@
 ﻿using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
 using System;
 using System.Collections.Generic;
 using System.Text;
