+++ src/Snowflake.Framework.Remoting.GraphQL/Model/Device/Mapped/DeviceLayoutMappingType.cs
@@ -1,6 +1,6 @@
 ﻿using HotChocolate.Types;
 using Snowflake.Remoting.GraphQL.Model.Device.Mapped;
-using Snowflake.Input.Controller.Mapped;
+using Snowflake.Input.Controller;
 using Snowflake.Input.Device;
 using System;
 using System.Collections.Generic;
