+++ src/Snowflake.Framework.Tests/Configuration/InputMapperTests.cs
@@ -8,7 +8,6 @@ using System.Threading.Tasks;
 using Snowflake.Configuration;
 using Snowflake.Configuration.Input;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
 using Snowflake.Input.Device;
 using Snowflake.Input.Tests;
 using Snowflake.Services;
