+++ src/Snowflake.Framework.Tests/Model/ControllerElementMappingsStoreTests.cs
@@ -6,7 +6,6 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
 using Snowflake.Input.Device;
 using Snowflake.Model.Database;
 using Snowflake.Model.Database.Models;
