+++ src/Snowflake.Framework/Model/Database/ControllerElementMappingProfileStore.Async.cs
@@ -4,8 +4,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
-using Snowflake.Input.Device;
 using Snowflake.Model.Database.Extensions;
 using Snowflake.Model.Database.Models;
