+++ src/Snowflake.Framework/Model/Database/ControllerElementMappingProfileStore.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
 using Snowflake.Input.Device;
 using Snowflake.Model.Database.Extensions;
 using Snowflake.Model.Database.Models;
