+++ src/Snowflake.Framework/Model/Database/Extensions/ControllerElementExtensions.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
-using Snowflake.Input.Device;
 using Snowflake.Model.Database.Models;
 
 namespace Snowflake.Model.Database.Extensions
