+++ src/Snowflake.Framework/Orchestration/Extensibility/EmulatedController.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Text;
 using Snowflake.Input.Controller;
-using Snowflake.Input.Controller.Mapped;
 using Snowflake.Input.Device;
 
 namespace Snowflake.Orchestration.Extensibility
